requirements.txt
asgiref==3.7.2
Django==4.2.6
django-jsonfield==1.4.1
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
numpy==1.26.0
pandas==2.1.1
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
